// apps/client/socket/socket.events.ts
export enum SOCKET_EVENT {
  CREATE_QUIZ_CSS = 'CREATE_QUIZ_CSS',
  CREATE_QUIZ_SSC = 'CREATE_QUIZ_SSC',

  JOIN_ROOM_CSS = 'JOIN_ROOM_CSS',
  JOIN_ROOM_SSC = 'JOIN_ROOM_SSC',

  LEAVE_ROOM_CSS = 'LEAVE_ROOM_CSS',
  LEAVE_ROOM_SSC = 'LEAVE_ROOM_SSC',

  STATISTICAL_ROOM_SSC = 'STATISTICAL_ROOM_SSC',

  START_QUIZ_CSS = 'START_QUIZ_CSS',
  START_QUIZ_SSC = 'START_QUIZ_SSC',

  ANSWER_THE_QUESTION_CSS = 'ANSWER_THE_QUESTION_CSS',
  ANSWER_THE_QUESTION_SSC = 'ANSWER_THE_QUESTION_SSC',

  TAKE_THE_QUESTION_CSS = 'TAKE_THE_QUESTION_CSS',
  TAKE_THE_QUESTION_SSC = 'TAKE_THE_QUESTION_SSC',
}
// answer the question
// take the question